fix: no-missing-keys rule reports false positive with trailing dot
For a translation key like $t('missing.') the parse function from core utilities returns undefined.

Previously this was turned into an empty array which caused all of those paths to be false positives.

If the given key can't be parsed as "path" properly, the correct handling is to treat it as a single item path with the provided string as the only item.

This also revealed another faulty test, where keypath="'hello'" was passing even though it should not.